Move variable interpolate command into bosh manifest package
[#165272723](https://www.pivotaltracker.com/story/show/165272723)
